requirements.txt
asgiref==3.4.1
base45==0.4.3
cachetools==4.2.4
cbor2==5.4.2
certifi==2021.10.8
charset-normalizer==2.0.7
ci-info==0.2.0
click==8.0.3
configobj==5.0.6
configparser==5.1.0
Django==3.2.9
etelemetry==0.2.2
filelock==3.3.2
fitz==0.0.1.dev2
flynn==1.0.0b2
future==0.18.2
google-auth==2.3.3
google-auth-oauthlib==0.4.6
gspread==4.0.1
httplib2==0.20.2
idna==3.3
isodate==0.6.0
lxml==4.6.4
networkx==2.6.3
nibabel==3.2.1
nipype==1.7.0
numpy==1.21.4
oauth2client==4.1.3
oauthlib==3.1.1
packaging==21.2
pandas==1.3.4
pathlib==1.0.1
Pillow==8.4.0
prov==2.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydot==1.4.2
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.3
pyxnat==1.4
pyzbar==0.1.8
rdflib==6.0.2
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.7.2
simplejson==3.17.5
six==1.16.0
sqlparse==0.4.2
traits==6.3.1
urllib3==1.26.7
